requirements.txt
flask==2.2.3
python-dotenv==1.0.0
werkzeug==2.2.3
requests==2.28.2
redis==4.5.1
PyYAML==6.0
gunicorn==20.1.0 